Split `dma` into multiple modules
